requirements.txt
albumentations==1.1.0
autopep8==1.6.0
brotlipy==0.7.0
certifi==2021.10.8
cffi @ file:///C:/ci_310/cffi_1642682485096/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cryptography @ file:///C:/ci/cryptography_1652101772384/work
cycler==0.11.0
einops==0.4.1
fonttools==4.33.3
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.19.2
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.2
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
networkx==2.8.1
numpy==1.22.3
opencv-python-headless==4.5.5.64
packaging==21.3
pandas==1.4.2
Pillow==8.1.2
pycodestyle==2.8.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pygame==2.1.2
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
scikit-image==0.19.2
scikit-learn==1.1.0
scipy==1.8.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
threadpoolctl==3.1.0
tifffile==2022.5.4
toml==0.10.2
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3 @ file:///C:/ci/urllib3_1650640043075/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
